autotest: autotestgenerator not static

+++ tsl2.nano.h5/src/main/de/tsl2/nano/h5/WebSecurity.java
@@ -45,7 +45,7 @@ public class WebSecurity {
     }
 	private String getAntiCSRFKey() {
 		if (antiCSRFKey == null)
-			antiCSRFKey = Crypt.generatePassword(16);
+			antiCSRFKey = Crypt.generatePassword((byte)16);
 		return antiCSRFKey;
 	}
 	public void checkAntiCSRFToken(NanoH5Session session, String token) {
